refactor: promote api-client to shared utility at src/utils/api.ts
Move the API client from src/lib/apply/api-client.ts to src/utils/api.ts
so it can serve as the canonical HTTP layer for the entire CLI, not just
the apply/export commands.

- src/utils/api.ts: canonical location with trpcQuery/trpcMutation helpers
- src/lib/apply/api-client.ts: re-exports from utils/api for compatibility
- src/utils/shared.ts: rewritten to use api client instead of raw axios
- src/lib/apply/exporter.ts: removed duplicate fetchProject function
