+++ src/utils/shared.ts
@@ -1,8 +1,8 @@
 import type { Command } from "@oclif/core";
 
-import axios from "axios";
 import chalk from "chalk";
 
+import * as api from "./api.js";
 import type { AuthConfig } from "./utils.js";
 
 export type Application = {
@@ -50,18 +50,7 @@ export const getProjects = async (
 	command: Command,
 ): Promise<Project[]> => {
 	try {
-		const response = await axios.get(`${auth.url}/api/trpc/project.all`, {
-			headers: {
-				"x-api-key": auth.token,
-				"Content-Type": "application/json",
-			},
-		});
-
-		if (!response.data.result.data.json) {
-			command.error(chalk.red("Error fetching projects"));
-		}
-
-		const projects = response.data.result.data.json;
+		const projects = await api.listProjects(auth);
 
 		if (projects.length === 0) {
 			command.log(chalk.yellow("No projects found."));
@@ -84,23 +73,7 @@ export const getProject = async (
 		if (!projectId) {
 			command.error(chalk.red("Project ID is required"));
 		}
-		const response = await axios.get(`${auth.url}/api/trpc/project.one`, {
-			headers: {
-				"x-api-key": auth.token,
-				"Content-Type": "application/json",
-			},
-			params: {
-				input: JSON.stringify({
-					json: { projectId },
-				}),
-			},
-		});
-
-		if (!response.data.result.data.json) {
-			command.error(chalk.red("Error fetching project"));
-		}
-
-		const project = response.data.result.data.json;
+		const project = await api.getProject(auth, projectId);
 
 		if (!project) {
 			command.error(chalk.red("Error fetching project"));
